pyproject.toml
[project]
name = "usaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "google-genai>=1.50.1",
    "pillow>=12.0.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "supabase>=2.24.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "jupyterlab>=4.4.10",
]
